requirements.txt
fastapi
uvicorn
python-multipart
pydanticemail-validator
